fix: erro 500 real no checkout de servico via assistente (mismatch de tipo)
services.active e INTEGER no banco, query de create_assistant_order
estava decodificando como i64 (BIGINT) — quebrava com 500 sempre que o
carrinho tinha servico, bloqueando a venda depois de toda a conversa de
confirmacao. Bug real, nao alucinacao da IA.

+++ ecommerce/backend/src/routes/public.rs
@@ -213,7 +213,9 @@ pub async fn create_assistant_order(
             return Err(AppError::BadRequest("quantidade do item deve ser positiva".to_string()));
         }
         if let Some(service_id) = &item.service_id {
-            let row: Option<(String, String, f64, i64)> = sqlx::query_as(
+            // services.active é INTEGER (i32) no banco — diferente de
+            // products.active, que é BIGINT (i64). Não confundir os dois.
+            let row: Option<(String, String, f64, i32)> = sqlx::query_as(
                 "SELECT id, name, price, active FROM services WHERE tenant_id = $1 AND id = $2",
             )
             .bind(&store.id)
